+++ app/components/legacy/extra/exit-confirmation-modal.hbs
