+++ app/components/legacy/form/formatted-triples.hbs
