+++ app/components/legacy/form/instance.hbs
@@ -29,7 +29,7 @@
             @skin="3"
             class="au-u-margin-bottom"
           >Content</AuHeading>
-          <Form::FormattedTriples @triples={{this.sourceTriples}} />
+          <Legacy::Form::FormattedTriples @triples={{this.sourceTriples}} />
         </section>
       </div>
     </div>
