+++ app/components/legacy/form/new-instance.hbs
@@ -40,7 +40,7 @@
             @skin="3"
             class="au-u-margin-bottom"
           >Content</AuHeading>
-          <Form::FormattedTriples @triples={{this.sourceTriples}} />
+          <Legacy::Form::FormattedTriples @triples={{this.sourceTriples}} />
         </section>
       </div>
     </div>
