+++ app/components/legacy/leidinggevendenbeheer/functionaris-form.hbs
@@ -1,10 +1,17 @@
 <div class="au-c-body-container au-c-body-container--scroll">
   <div class="au-o-box">
-    <Leidinggevendenbeheer::PersoonCard @persoon={{@model.isBestuurlijkeAliasVan}} @hideHeader={{true}} />
+    <Legacy::Leidinggevendenbeheer::PersoonCard
+      @persoon={{@model.isBestuurlijkeAliasVan}}
+      @hideHeader={{true}}
+    />
 
     <AuHr @size="large" />
 
-    <AuHeading @level="2" @skin="4" class="au-u-margin-bottom-small">Functie</AuHeading>
+    <AuHeading
+      @level="2"
+      @skin="4"
+      class="au-u-margin-bottom-small"
+    >Functie</AuHeading>
     <div class="au-o-grid au-o-grid--tiny">
       <div class="au-o-grid__item au-u-1-2 au-u-1-3@medium">
         <AuLabel for="date-picker-start">Start aanstellingsperiode</AuLabel>
@@ -15,7 +22,7 @@
           id="date-picker-start"
         />
         {{#unless @model.start}}
-        <AuHelpText @error="true">Start aanstellingsperiode is verplicht</AuHelpText>
+          <AuHelpText @error="true">Start aanstellingsperiode is verplicht</AuHelpText>
         {{/unless}}
       </div>
       <div class="au-o-grid__item au-u-1-2 au-u-1-3@medium">
@@ -29,21 +36,21 @@
       </div>
     </div>
     {{#if (gt this.statusOptions.length 1)}}
-    {{!-- TODO: Use a fieldset + legend combo: https://github.com/appuniversum/ember-appuniversum/issues/210 --}}
-    <AuLabel class="au-u-margin-top">Status</AuLabel>
-    <div class="au-o-grid au-o-grid--tiny">
-      {{#each this.statusOptions as |status|}}
-      <div class="au-o-grid__item au-u-1-2 au-u-1-4@medium">
-        <AuControlRadio
-          @label={{status.label}}
-          @name="functionary-status"
-          @value={{status.label}}
-          @onChange={{fn (mut @model.status) status}}
-          checked={{eq status.id @model.status.id}}
-        />
+      {{! TODO: Use a fieldset + legend combo: https://github.com/appuniversum/ember-appuniversum/issues/210 }}
+      <AuLabel class="au-u-margin-top">Status</AuLabel>
+      <div class="au-o-grid au-o-grid--tiny">
+        {{#each this.statusOptions as |status|}}
+          <div class="au-o-grid__item au-u-1-2 au-u-1-4@medium">
+            <AuControlRadio
+              @label={{status.label}}
+              @name="functionary-status"
+              @value={{status.label}}
+              @onChange={{fn (mut @model.status) status}}
+              checked={{eq status.id @model.status.id}}
+            />
+          </div>
+        {{/each}}
       </div>
-      {{/each}}
-    </div>
     {{/if}}
   </div>
 </div>
