+++ app/components/legacy/persoon/create-persoon.hbs
@@ -87,51 +87,53 @@
         {{/let}}
       </div>
       {{#if this.isNationalityFieldRequired}}
-      {{#let (if this.errors.nationaliteit true false) as |showError|}}
-        <div>
-          <AuLabel for="person-nationalities"
-            @error={{showError}}
-            @required={{true}}
+        {{#let (if this.errors.nationaliteit true false) as |showError|}}
+          <div>
+            <AuLabel
+              for="person-nationalities"
+              @error={{showError}}
+              @required={{true}}
             >
-            Nationaliteit
-          </AuLabel>
+              Nationaliteit
+            </AuLabel>
 
-          <Shared::Persoon::NationalitySelector
-            id="person-nationalities"
-            @allowClear={{not this.isNationalityFieldRequired}}
-            @selected={{this.nationaliteit}}
-            @onChange={{fn (mut this.nationaliteit)}}
-            @error={{showError}}
-          />
-        </div>
-        {{#if showError}}
-          <AuHelpText @error="true">
-            {{this.errors.nationaliteit}}
-          </AuHelpText>
-        {{/if}}
-      {{/let}}
+            <Legacy::Persoon::NationalitySelector
+              id="person-nationalities"
+              @allowClear={{not this.isNationalityFieldRequired}}
+              @selected={{this.nationaliteit}}
+              @onChange={{fn (mut this.nationaliteit)}}
+              @error={{showError}}
+            />
+          </div>
+          {{#if showError}}
+            <AuHelpText @error="true">
+              {{this.errors.nationaliteit}}
+            </AuHelpText>
+          {{/if}}
+        {{/let}}
       {{/if}}
       <div>
         {{#let (if this.errors.geboorte true false) as |showError|}}
-        <AuLabel for="mandataris-geboorte"
-          @required={{this.isGeboorteFieldRequired}}
-          @error={{showError}}
+          <AuLabel
+            for="mandataris-geboorte"
+            @required={{this.isGeboorteFieldRequired}}
+            @error={{showError}}
           >
-          Geboortedatum
-        </AuLabel>
-        <AuDateInput
-          @value={{this.birthDate}}
-          @onChange={{this.setDateOfBirth}}
-          @error={{showError}}
-          @width="block"
-          id="mandataris-geboorte"
-        />
-        {{#if showError}}
-          <AuHelpText @error="true">
-            {{this.errors.geboorte}}
-          </AuHelpText>
-        {{/if}}
-      {{/let}}
+            Geboortedatum
+          </AuLabel>
+          <AuDateInput
+            @value={{this.birthDate}}
+            @onChange={{this.setDateOfBirth}}
+            @error={{showError}}
+            @width="block"
+            id="mandataris-geboorte"
+          />
+          {{#if showError}}
+            <AuHelpText @error="true">
+              {{this.errors.geboorte}}
+            </AuHelpText>
+          {{/if}}
+        {{/let}}
       </div>
       <div>
         <AuLabel @error={{if this.errors.geslacht true}} @required={{true}}>
