+++ app/templates/application.hbs
@@ -36,7 +36,7 @@
           </AuLink>
         </AuDropdown>
       {{else}}
-        <LoginButton @isCompact={{true}} />
+        <Shared::LoginButton @isCompact={{true}} />
       {{/if}}
     </li>
   </AuMainHeader>
