+++ app/templates/auth/callback-loading.hbs
@@ -1,5 +1,5 @@
 {{page-title "Aan het aanmelden"}}
 
-<PageLoader>
+<Shared::PageLoader>
   Aan het aanmelden
-</PageLoader>
+</Shared::PageLoader>
