+++ app/templates/auth/login.hbs
@@ -1,5 +1,5 @@
 {{page-title "Aan het doorverwijzen"}}
 
-<PageLoader>
+<Shared::PageLoader>
   Aan het doorverwijzen
-</PageLoader>
+</Shared::PageLoader>
