+++ app/templates/auth/logout-loading.hbs
@@ -1,5 +1,5 @@
 {{page-title "Aan het afmelden"}}
 
-<PageLoader>
+<Shared::PageLoader>
   Aan het afmelden
-</PageLoader>
+</Shared::PageLoader>
