+++ app/templates/auth/switch.hbs
@@ -1,5 +1,5 @@
 {{page-title "Aan het doorverwijzen"}}
 
-<PageLoader>
+<Shared::PageLoader>
   Aan het het doorverwijzen
-</PageLoader>
+</Shared::PageLoader>
