+++ app/templates/contact.hbs
@@ -19,7 +19,7 @@
     <section class="au-o-region-large">
       <div class="au-o-layout">
         <AuHeading class="au-u-margin-bottom">Contact</AuHeading>
-        <ContactForm />
+        <Legacy::Extra::ContactForm />
       </div>
     </section>
