+++ app/templates/legacy/formbeheer/form/instance.hbs
@@ -1,5 +1,5 @@
 {{page-title "Instance"}}
-<Form::Instance
+<Legacy::Form::Instance
   @instanceId={{this.model.instanceId}}
   @form={{this.model.form}}
   @onSave={{this.onSave}}
