+++ app/templates/legacy/formbeheer/form/instances.hbs
@@ -1,5 +1,5 @@
 {{page-title "Instances"}}
-<Form::InstanceTable
+<Legacy::Form::InstanceTable
   @onCreate={{this.onCreate}}
   @form={{this.model.form}}
   @onRemoveInstance={{this.onRemoveInstance}}
