+++ app/templates/legacy/formbeheer/form/new.hbs
@@ -1,2 +1,5 @@
 {{page-title "New"}}
-<Form::NewInstance @onCreate={{this.onCreate}} @form={{this.model.form}} />
+<Legacy::Form::NewInstance
+  @onCreate={{this.onCreate}}
+  @form={{this.model.form}}
+/>
