+++ app/templates/legacy/leidinggevendenbeheer/bestuursfuncties/bestuursfunctie/functionarissen/index.hbs
@@ -56,7 +56,7 @@
   </Group>
 </AuToolbar>
 
-<Leidinggevendenbeheer::FunctionarisTable
+<Legacy::Leidinggevendenbeheer::FunctionarisTable
   class="au-c-body-container"
   @content={{this.model}}
   @sort={{this.sort}}
