+++ app/templates/legacy/leidinggevendenbeheer/bestuursfuncties/bestuursfunctie/functionarissen/new/index.hbs
@@ -7,12 +7,12 @@
   </Group>
 </AuToolbar>
 
-<Shared::Persoon::PersoonSearchForm
+<Legacy::Persoon::PersoonSearchForm
   @onSelect={{this.choosePeriode}}
   @onCreateNewPerson={{this.createNewPerson}}
   @onCancel={{this.cancel}}
   @showDefaultHead={{false}}
   class="au-c-body-container"
-  >
+>
   <AuHeading @skin="2">Voeg nieuwe persoon toe</AuHeading>
-</Shared::Persoon::PersoonSearchForm>
+</Legacy::Persoon::PersoonSearchForm>
