+++ app/templates/legacy/leidinggevendenbeheer/bestuursfuncties/bestuursfunctie/functionarissen/new/periode.hbs
@@ -4,28 +4,46 @@
   </Group>
   <Group class="au-c-toolbar__group--row au-u-margin-top-none">
     <AuHeading @level="3" @skin="2">
-      2. Voeg aanstellingsperiode voor {{this.model.isBestuurlijkeAliasVan.gebruikteVoornaam}}
-      {{this.model.isBestuurlijkeAliasVan.achternaam}} als {{this.model.bekleedt.rol.label}} toe</AuHeading>
+      2. Voeg aanstellingsperiode voor
+      {{this.model.isBestuurlijkeAliasVan.gebruikteVoornaam}}
+      {{this.model.isBestuurlijkeAliasVan.achternaam}}
+      als
+      {{this.model.bekleedt.rol.label}}
+      toe</AuHeading>
   </Group>
 </AuToolbar>
 
-<Leidinggevendenbeheer::FunctionarisForm @model={{this.model}} as |isValid|>
+<Legacy::Leidinggevendenbeheer::FunctionarisForm
+  @model={{this.model}}
+  as |isValid|
+>
   <AuToolbar @border="top" @size="large" as |Group|>
     <Group>
-      <AuButton {{on 'click' this.goBackToSearch}} @skin="tertiary" @disabled={{this.save.isRunning}}>
+      <AuButton
+        {{on "click" this.goBackToSearch}}
+        @skin="tertiary"
+        @disabled={{this.save.isRunning}}
+      >
         <AuIcon @icon="arrow-left" @alignment="left" />
         Vorige stap
       </AuButton>
     </Group>
     <Group>
       <AuButtonGroup>
-        <AuButton @skin="secondary" @disabled={{this.save.isRunning}} {{on 'click' this.cancel}}>
+        <AuButton
+          @skin="secondary"
+          @disabled={{this.save.isRunning}}
+          {{on "click" this.cancel}}
+        >
           Annuleer
         </AuButton>
-        <AuButton @disabled={{if (or (not isValid) this.save.isRunning) "true" }} {{on "click" (perform this.save)}}>
+        <AuButton
+          @disabled={{if (or (not isValid) this.save.isRunning) "true"}}
+          {{on "click" (perform this.save)}}
+        >
           Voeg aanstellingsperiode toe
         </AuButton>
       </AuButtonGroup>
     </Group>
   </AuToolbar>
-</Leidinggevendenbeheer::FunctionarisForm>
+</Legacy::Leidinggevendenbeheer::FunctionarisForm>
