+++ app/templates/legacy/mandatenbeheer/fracties-new/new.hbs
@@ -1,5 +1,5 @@
 {{page-title "New"}}
-<Form::NewInstance
+<Legacy::Form::NewInstance
   @onCreate={{this.onCreate}}
   @form={{this.model.form}}
   @buildSourceTtl={{this.buildSourceTtl}}
