+++ app/templates/legacy/mandatenbeheer/mandatarissen/new-person.hbs
@@ -4,9 +4,8 @@
   </Group>
 </AuToolbar>
 
-<Shared::Persoon::CreatePersoon
+<Legacy::Persoon::CreatePersoon
   @prefilledValues={{this.personPrefilledValues}}
   @onCreate={{this.onCreate}}
   @onCancel={{this.onCancel}}
 />
-
