+++ app/templates/legacy/mandatenbeheer/mandatarissen/new.hbs
@@ -1,9 +1,8 @@
-<Shared::Persoon::PersoonSearchForm
+<Legacy::Persoon::PersoonSearchForm
   class="au-c-body-container"
   @onCreateNewPerson={{this.createNewPerson}}
   @onCancel={{this.cancel}}
   @onSelect={{this.selectPersoon}}
-  >
- <AuHeading @skin="2">Voeg mandataris toe uit de lijst van verkozenen</AuHeading>
-</Shared::Persoon::PersoonSearchForm>
-
+>
+  <AuHeading @skin="2">Voeg mandataris toe uit de lijst van verkozenen</AuHeading>
+</Legacy::Persoon::PersoonSearchForm>
