+++ app/templates/login.hbs
@@ -40,7 +40,7 @@
             <div
               class="au-o-grid__item au-u-1-4@small au-u-1-2@medium au-u-text-right@medium"
             >
-              <LoginButton />
+              <Shared::LoginButton />
             </div>
           </div>
         </div>
